Fix for Custom Potential Bug
There was an issue where the numAdsorbates being fed to the program being different than the numadsorbates in the custom potential file resulted in very strange paths being generated. A fix by checking how many adsorbates are in the custom potential file and reassigning the numAdsorbates value has been made.